fix(factories): Event serializer need a place; factories had to be splitted to avoid circular import
